+++ src/library/BarChart/BarSegment.tsx
