+++ src/library/BarChart/LegendItem.tsx
