+++ src/library/BarChart/defaults.ts
